Move ER-specific deployment tests
